GCA-167 create workloadService, CalculateService and tests

+++ trainers-workload-service/src/main/java/com/trainersworkloadservice/model/dto/WorkloadEventRequest.java
@@ -13,7 +13,7 @@ public record WorkloadEventRequest(
         @NotBlank String lastName,
         @NotBlank Boolean isActive,
         @NotNull @JsonFormat(pattern = "yyyy-MM-dd") LocalDate trainingDate,
-        @Positive int trainingDuration,
+        @Positive Long trainingDuration,
         @NotNull ActionType actionType
 ) {
 }
